TrueTypeFont[SubSet]: Support 0/3 cmap
Read support for it in TrueTypeFont is trivial, since it is semantically
equivalent to a 3/1 cmap. However, the PDF specification doesn't mention
that it has to be supported, so many PDF readers don't. To solve the
issue, TrueTypeFontSubSet can write a new (3/1 format 4) cmap table, if
requested. This doesn't cover the CFF case, though.