TrueTypeFontSubSet: Create both 3/1 and 3/10 cmap subtables
Writing a 3/1 format 12.0 cmap subtable is not a good idea. While Adobe
Reader and Okular display the result just fine, other readers (Mac OS X
Preview) don't. According to the OpenType specification a 3/1 subtable
should use format 4 anyway.

Now we always write a 3/1 format 4 table. If we have non-BMP chars or
the subtable gets too big (> 64 KiB), we omit characters, but still
write the subtable. In that case we additionally write a 3/10 format
12 subtable.